auth is working

+++ back/src/main/java/com/alten/shop/ProductsController.java
@@ -1,10 +1,17 @@
 package com.alten.shop;
 
 import java.util.List;
+import java.util.Objects;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
+@CrossOrigin
 @RestController
 @RequestMapping("/api/products")
 public class ProductsController {
@@ -17,21 +24,33 @@ public class ProductsController {
     }
 
     @PostMapping("")
-    public String create(@RequestBody Product product) {
-        productService.create(product);
-        return "OK";
+    public ResponseEntity<String> create(@RequestBody Product product) {
+        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (Objects.equals(currentUser.getEmail(), "admin@admin.com")) {
+            productService.create(product);
+            return ResponseEntity.ok("OK");
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
     @PatchMapping("/{productId}")
-    public String update(@PathVariable(value="productId") String productId, @RequestBody Product product) {
-        product.setId(Long.parseLong(productId));
-        productService.update(product);
-        return "OK";
+    public ResponseEntity<String> update(@PathVariable(value="productId") String productId, @RequestBody Product product) {
+        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (Objects.equals(currentUser.getEmail(), "admin@admin.com")) {
+            product.setId(Long.parseLong(productId));
+            productService.update(product);
+            return ResponseEntity.ok("OK");
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
     @DeleteMapping("/{productId}")
-    public String delete(@PathVariable(value="productId") String productId) {
-        productService.delete(Long.parseLong(productId));
-        return "OK";
+    public ResponseEntity<String> delete(@PathVariable(value="productId") String productId) {
+        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (Objects.equals(currentUser.getEmail(), "admin@admin.com")) {
+            productService.delete(Long.parseLong(productId));
+            return ResponseEntity.ok("OK");
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 }
